chore(jans-cedarling): rename `models` module to `common`
