+++ jans-cedarling/cedarling/src/common/cedar_schema.rs
@@ -78,7 +78,7 @@ mod deserialize {
     #[cfg(test)]
     mod tests {
         use super::*;
-        use crate::models::policy_store::PolicyStoreMap;
+        use crate::common::policy_store::PolicyStoreMap;
 
         #[test]
         fn test_read_ok() {
