+++ jans-cedarling/cedarling/src/init/policy_store.rs
@@ -6,7 +6,7 @@
  */
 
 use crate::bootstrap_config::policy_store_config::{PolicyStoreConfig, PolicyStoreSource};
-use crate::models::policy_store::{PolicyStore, PolicyStoreMap};
+use crate::common::policy_store::{PolicyStore, PolicyStoreMap};
 
 /// Error cases for loading policy
 #[derive(Debug, thiserror::Error)]
